fix architecture update & add-table
