+++ Cargo.toml
@@ -20,6 +20,7 @@ thiserror = "1.0.50"
 validator = { version = "0.12", features = ["derive"] }
 sqlx = { version = "0.5.13", features = ["postgres", "any", "runtime-tokio-rustls"] }
 dotenv = "0.15.0"
+tower-http = { version = "0.4.4", features = ["cors"] }
 
 [features]
 default = ["database-test"]
