+++ src/handlers/todo.rs
@@ -1,0 +1,89 @@
+use crate::repositories::todo::{CreateTodo, TodoRepository, UpdateTodo};
+use axum::{
+    async_trait,
+    extract::{Extension, FromRequest, Path},
+    http::StatusCode,
+    response::IntoResponse,
+    Json, RequestExt,
+};
+use hyper::Request;
+use std::sync::Arc;
+use validator::Validate;
+
+#[derive(Debug)]
+pub struct ValidatedJson<T>(pub T);
+
+#[async_trait]
+impl<S, B, T> FromRequest<S, B> for ValidatedJson<T>
+where
+    B: Send + 'static,
+    S: Send + Sync,
+    T: Validate + 'static,
+    Json<T>: FromRequest<(), B>,
+{
+    type Rejection = (StatusCode, &'static str);
+
+    // https://dev.to/ayush1325/validating-json-request-in-axum-2n34
+    async fn from_request(req: Request<B>, _state: &S) -> Result<Self, Self::Rejection> {
+        let Json(data) = req
+            .extract::<Json<T>, _>()
+            .await
+            .map_err(|_| (StatusCode::BAD_REQUEST, "Invalid JSON body"))?;
+        data.validate()
+            .map_err(|_| (StatusCode::BAD_REQUEST, "Invalid JSON body"))?;
+        Ok(Self(data))
+    }
+}
+
+// **memo 1**
+// https://qiita.com/Sicut_study/items/5e5d6cce5ba48c225367
+pub async fn create_todo<T: TodoRepository>(
+    // この並びが逆だとコンパイルエラーになる
+    Extension(repository): Extension<Arc<T>>,
+    ValidatedJson(payload): ValidatedJson<CreateTodo>,
+) -> Result<impl IntoResponse, StatusCode> {
+    let todo = repository
+        .create(payload)
+        .await
+        .or(Err(StatusCode::NOT_FOUND))?;
+
+    Ok((StatusCode::CREATED, Json(todo)))
+}
+
+pub async fn find_todo<T: TodoRepository>(
+    Path(id): Path<i32>,
+    Extension(repository): Extension<Arc<T>>,
+) -> Result<impl IntoResponse, StatusCode> {
+    let todo = repository.find(id).await.or(Err(StatusCode::NOT_FOUND))?;
+    Ok((StatusCode::OK, Json(todo)))
+}
+
+pub async fn all_todo<T: TodoRepository>(
+    Extension(repository): Extension<Arc<T>>,
+) -> Result<impl IntoResponse, StatusCode> {
+    let todo = repository.all().await.unwrap();
+    Ok((StatusCode::OK, Json(todo)))
+}
+
+pub async fn update_todo<T: TodoRepository>(
+    Extension(repository): Extension<Arc<T>>,
+    Path(id): Path<i32>,
+    ValidatedJson(payload): ValidatedJson<UpdateTodo>,
+) -> Result<impl IntoResponse, StatusCode> {
+    let todo = repository
+        .update(id, payload)
+        .await
+        .or(Err(StatusCode::NOT_FOUND))?;
+    Ok((StatusCode::CREATED, Json(todo)))
+}
+
+pub async fn delete_todo<T: TodoRepository>(
+    Path(id): Path<i32>,
+    Extension(repository): Extension<Arc<T>>,
+) -> StatusCode {
+    repository
+        .delete(id)
+        .await
+        .map(|_| StatusCode::NO_CONTENT)
+        .unwrap_or(StatusCode::NOT_FOUND)
+}
